+++ source/components/AppHeader.js
@@ -1,4 +1,4 @@
-import { normalize } from '../js/utils.js';
+import { normalize } from '../js/css.js';
 import { navigateTo } from '../js/navigation.js';
 import { isMuted, toggleMute, switchStop } from '../js/audio.js';
