+++ source/tests/randomHouse.test.js
@@ -1,0 +1,8 @@
+import { randomHouse, houses } from '../js/utils.js';
+
+describe('randomHouse', () => {
+  test('should return a random house from the houses array', () => {
+    const house = randomHouse();
+    expect(houses).toContain(house);
+  });
+});
